Add changes from code review: * Move connectivity.py to utils (https://github.com/mosdef-hub/gmso/pull/356#issuecomment-601437370) * Test n_dihedrals and n_impropers in unit tests (https://github.com/mosdef-hub/gmso/pull/356#issuecomment-601658319) * Ensure that NoneType matches are ignored
